refactor: domain 레이어 서브패키지 분리 (account/codef/manual/summary)
