+++ src/main/java/com/team/independence/asset/domain/account/CardAccount.java
@@ -1,4 +1,4 @@
-package com.team.independence.asset.domain;
+package com.team.independence.asset.domain.account;
 
 import lombok.*;
