+++ src/main/java/com/team/independence/asset/domain/codef/Institution.java
@@ -1,4 +1,4 @@
-package com.team.independence.asset.domain;
+package com.team.independence.asset.domain.codef;
 
 import lombok.*;
