+++ src/main/java/com/team/independence/asset/domain/manual/ManualAsset.java
@@ -1,4 +1,4 @@
-package com.team.independence.asset.domain;
+package com.team.independence.asset.domain.manual;
 
 import lombok.*;
 
@@ -9,13 +9,11 @@ import java.time.LocalDateTime;
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
-public class AssetSummary {
+public class ManualAsset {
     private Long id;
     private Long memberId;
-    private Long totalAssets;
-    private Long loanBalance;
-    private Long monthlySavings;
-    private LocalDateTime syncedAt;
+    private String assetType;  // DEPOSIT: 현재 거주 보증금
+    private Long amount;
     private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
 }
