+++ src/main/java/com/team/independence/asset/domain/summary/AssetSummary.java
@@ -1,4 +1,4 @@
-package com.team.independence.asset.domain;
+package com.team.independence.asset.domain.summary;
 
 import lombok.*;
 
@@ -9,11 +9,13 @@ import java.time.LocalDateTime;
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
-public class ManualAsset {
+public class AssetSummary {
     private Long id;
     private Long memberId;
-    private String assetType;  // DEPOSIT: 현재 거주 보증금
-    private Long amount;
+    private Long totalAssets;
+    private Long loanBalance;
+    private Long monthlySavings;
+    private LocalDateTime syncedAt;
     private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
 }
